requirements.txt
# Core Libraries
fastapi==0.104.1
uvicorn[standard]==0.24.0
librosa==0.10.1
numpy==1.24.3
pillow==10.1.0
torch==2.1.1
torchvision==0.16.1
diffusers==0.27.2
transformers==4.40.0
accelerate==0.25.0
scipy==1.11.4
huggingface-hub==0.23.0
matplotlib==3.8.2
aiofiles==23.2.1

# Speech Recognition (using git+https to avoid build issues)
git+https://github.com/openai/whisper.git

# Testing Libraries
pytest==7.4.3
pytest-asyncio==0.21.1
hypothesis==6.92.1
httpx==0.25.2
pytest-cov==4.1.0

# Additional Dependencies
python-multipart==0.0.6
